Fixed camera jitter.

+++ ASCIIWorld/src/asciiWorld/Camera.java
@@ -82,7 +82,8 @@ public class Camera implements IHasPosition, IHasBounds, IHasRangeOfVision {
 		Vector3f position = getPosition();
 		
 		g.scale(scale, scale);
-		g.translate(_scaledCenter.x - position.x, _scaledCenter.y - position.y);
+		//g.translate(_scaledCenter.x - position.x, _scaledCenter.y - position.y);
+		g.translate((int)(_scaledCenter.x - position.x), (int)(_scaledCenter.y - position.y));
 	}
 	
 	public void reset(Graphics g) {
